Putting emulation in a separate file
