+++ drivers/pn532/Kconfig
@@ -22,11 +22,13 @@ config PN532_INIT_PRIORITY
     help
       PN532 device drivers init priority.
 
-config PN532_TRANSPORT_MOCK
-    bool "Use mock transport for PN532"
-    default n
-    help
-      Use a dummy transport backend for testing purposes.
+config EMUL_PN532
+  bool "Emulate a PN532 module"
+  default y
+  depends on EMUL
+  depends on PN532
+  help
+    It provides emulated I2C for testing purposes.
 
 module = PN532
 module-str = pn532
