+++ drivers/pn532/pn532.c
@@ -7,9 +7,7 @@
 
 #include <zephyr/device.h>
 #include <zephyr/devicetree.h>
-#include <zephyr/drivers/emul.h>
 #include <zephyr/drivers/i2c.h>
-#include <zephyr/drivers/i2c_emul.h>
 #include <zephyr/kernel.h>
 #include <zephyr/logging/log.h>
 LOG_MODULE_REGISTER(pn532, CONFIG_PN532_LOG_LEVEL);
@@ -19,10 +17,6 @@ LOG_MODULE_REGISTER(pn532, CONFIG_PN532_LOG_LEVEL);
 
 struct pn532_data {
     uint32_t dummy;
-};
-
-struct pn532_emul_cfg {
-    uint16_t addr;
 };
 
 struct pn532_config {
@@ -44,36 +38,6 @@ static DEVICE_API(pn532, pn532_api) = {
     .pn532_get_firmware_version = &get_firmware_version,
 };
 
-static int pn532_transfer_i2c(const struct emul *target, struct i2c_msg *msgs,
-				       int num_msgs, int addr)
-{
-	ARG_UNUSED(target);
-	ARG_UNUSED(msgs);
-	ARG_UNUSED(num_msgs);
-	ARG_UNUSED(addr);
-
-	return 0;
-}
-
-static const struct i2c_emul_api pn532_emul_api_i2c = {
-	.transfer = pn532_transfer_i2c,
-};
-
-/**
- * Set up a new emulator (I2C)
- *
- * @param emul Emulation information
- * @param parent Device to emulate
- * @return 0 indicating success (always)
- */
-static int emul_pn532_init(const struct emul *target, const struct device *parent)
-{
-	ARG_UNUSED(target);
-	ARG_UNUSED(parent);
-
-	return 0;
-}
-
 static int pn532_init(const struct device *dev)
 {
     int ret = pn532_transport_init(dev);
@@ -87,26 +51,20 @@ static int pn532_init(const struct device *dev)
     return 0;
 }
 
-#define PN532_EMUL(n)                                         \
-    static const struct pn532_emul_cfg pn532_emul_cfg_##n = { \
-        .addr = DT_INST_REG_ADDR(n),                          \
-    };                                                           \
-                                                                 \
-    EMUL_DT_INST_DEFINE(n, emul_pn532_init, NULL,           \
-                        &pn532_emul_cfg_##n, &pn532_emul_api_i2c, NULL)
-
-DT_INST_FOREACH_STATUS_OKAY(PN532_EMUL)
-
-#define PN532_DEFINE(inst)                                         \
-    static struct pn532_data data##inst;                           \
-                                                                   \
-    static const struct pn532_config config##inst = {              \
-        .i2c_dev = DEVICE_DT_GET(DT_INST_BUS(inst)),               \
-    };                                                             \
-                                                                   \
-    DEVICE_DT_INST_DEFINE(inst, pn532_init, NULL,                  \
-                          &data##inst, &config##inst,              \
-                          POST_KERNEL, CONFIG_PN532_INIT_PRIORITY, \
+#define PN532_DEFINE(inst)                            \
+    static struct pn532_data data##inst;              \
+                                                      \
+    static const struct pn532_config config##inst = { \
+        .i2c_dev = DEVICE_DT_GET(DT_INST_BUS(inst)),  \
+    };                                                \
+                                                      \
+    DEVICE_DT_INST_DEFINE(inst,                       \
+                          pn532_init,                 \
+                          NULL,                       \
+                          &data##inst,                \
+                          &config##inst,              \
+                          POST_KERNEL,                \
+                          CONFIG_PN532_INIT_PRIORITY, \
                           &pn532_api);
 
 DT_INST_FOREACH_STATUS_OKAY(PN532_DEFINE)
