+++ tests/drivers/pn532/prj.conf
@@ -9,4 +9,3 @@ CONFIG_ZTEST=y
 CONFIG_I2C=y
 CONFIG_PN532=y
 CONFIG_PN532_LOG_LEVEL=4
-CONFIG_PN532_TRANSPORT_MOCK=n
