+++ tests/drivers/pn532/testcase.yaml
@@ -11,4 +11,4 @@ tests:
       - drivers
       - pn532
     extra_configs:
-      - CONFIG_PN532_TRANSPORT_MOCK=n
+      - CONFIG_EMUL_PN532=y
